% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ABM_Field.R
\name{ABM_Field}
\alias{ABM_Field}
\title{Construct an ABM_Field object (internal)}
\usage{
ABM_Field(
  x,
  name,
  category = c("state", "active_state", "act_FUN", "stop_FUN", "plot_FUN", "report_FUN")
)
}
\arguments{
\item{x}{A value or function.}

\item{name}{Field name (single character string).}

\item{category}{One of \code{"state"}, \code{"active_state"}, \code{"act_FUN"},
\code{"stop_FUN"}, \code{"report_FUN"}, or \code{"plot_FUN"}. See
\link{Field} for the constructor that produces each category.}
}
\value{
An object of class \code{"ABM_Field"}.
}
\description{
Construct an ABM_Field object (internal)
}
\keyword{internal}
