% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ABM_Zip.R
\name{ABM_Zip}
\alias{ABM_Zip}
\title{Construct an ABM_Zip object (internal)}
\usage{
ABM_Zip(...)
}
\arguments{
\item{...}{Objects to be bundled.}
}
\value{
An object of class \code{"ABM_Zip"} (a list).
}
\description{
Internal low-level constructor for objects of class \code{"ABM_Zip"}.
}
\keyword{internal}
