% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-Game.R
\name{as.Game}
\alias{as.Game}
\title{Coerce an object to an ABM game}
\usage{
as.Game(x, ...)
}
\arguments{
\item{x}{An object to be coerced into an \code{\link{ABM_Game}}.}

\item{...}{Reserved for future extensions.}
}
\value{
An \code{\link{ABM_Game}} object.
}
\description{
\code{as.Game()} coerces an object into an \code{\link{ABM_Game}} object.
If \code{x} is already an \code{\link{ABM_Game}}, it is returned unchanged.
Otherwise, \code{x} is treated as \code{"stage"} and a new game object is created.
}
\seealso{
\code{\link[=Game]{Game()}}, \code{\link{ABM_Game}}
}
