% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_gpt_helper.R
\name{ask_gpt_helper}
\alias{ask_gpt_helper}
\title{Helper Functions for ask_gpt}
\description{
A set of internal functions used by the \code{ask_gpt} workflow for RAG integration.
These include functions for reading remote files, cleaning code, and transforming
roxygen2 documentation into plain text.
}
\keyword{internal}
