% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_obj.R
\name{copy_obj}
\alias{copy_obj}
\alias{copy_obj.ABM_Game}
\alias{copy_obj.R6}
\title{Deep copy ABM and R6 objects}
\usage{
copy_obj(x)

\method{copy_obj}{ABM_Game}(x)

\method{copy_obj}{R6}(x)
}
\arguments{
\item{x}{An object to copy. Supported classes are:
\itemize{
\item \code{ABM_Game}
\item \code{R6} (any other R6 object)
}}
}
\value{
A deep-cloned object of the same class as \code{x}.
}
\description{
Create a deep copy of an \code{ABM_Game} or other \code{R6} object.
}
\details{
This function provides a unified interface for cloning objects used in
rABM, dispatching on the class of \code{x}.

For \code{ABM_Game} objects, \code{copy_obj()} calls
\code{x$clone(deep = TRUE)} and then \code{$.rebind_dynamic_fields()} on
the clone, since R6's default \code{clone()} does not correctly carry
over the dynamically added methods and active bindings (\code{act_FUN},
\code{stop_FUN}, \code{report_FUN}, \code{plot_FUN}, \code{active_state})
that \code{ABM_Game} attaches at runtime (see \code{ABM_Game}).

For any other \code{R6} object, \code{copy_obj()} simply calls
\code{x$clone(deep = TRUE)}.

Objects that do not belong to the supported classes are not handled by
this function and will result in an error.
}
\examples{
pop <- 100
reproduce <- function() { self$pop <- self$pop * 1.1 }

G  <- Game(State(pop), Act(reproduce))
G2 <- copy_obj(G)
}
