% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ABM_Field.R
\name{.coerce_call_to_FUN}
\alias{.coerce_call_to_FUN}
\title{Coerce a call like fun(a=3) into a function with updated defaults (internal)}
\usage{
.coerce_call_to_FUN(x_call, envir = parent.frame())
}
\arguments{
\item{x_call}{A call object (e.g., \code{fun(a = 3)}).}

\item{envir}{Environment used to evaluate the call head (function object).}
}
\value{
A list with elements \code{fun} and \code{name_from_call}.
}
\description{
Coerce a call like fun(a=3) into a function with updated defaults (internal)
}
\keyword{internal}
