% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-field_typeof.R
\name{.field_typeof}
\alias{.field_typeof}
\title{Classify a field value by a coarse type label}
\usage{
.field_typeof(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
A length-1 character label.
}
\description{
Internal helper used for field registries and printing.
}
\keyword{internal}
