% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-format_input_R6.R
\name{.format_input_R6}
\alias{.format_input_R6}
\title{Validate and normalize input for R6 initialization}
\usage{
.format_input_R6(x)
}
\arguments{
\item{x}{An input object expected to be a named list.}
}
\value{
A named list.
}
\description{
Internal helper used in R6 \code{initialize()} methods to validate inputs.
}
\keyword{internal}
