% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_agent_helper.R
\name{.shape_agent_attr}
\alias{.shape_agent_attr}
\title{Format agent attribute into data.frame}
\usage{
.shape_agent_attr(attr, attr_sbs)
}
\arguments{
\item{attr}{A vector or data.frame representing an attribute.}

\item{attr_sbs}{Symbol of the attribute (from substitute()).}
}
\value{
A data.frame or NULL.
}
\description{
This function standardizes agent attributes passed in as either vector or data.frame
and returns a data.frame with a single column named appropriately.
}
\keyword{internal}
