% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ABM_Field.R
\name{.validate_name1}
\alias{.validate_name1}
\title{Validate a field name (internal)}
\usage{
.validate_name1(name)
}
\arguments{
\item{name}{A candidate field name.}
}
\value{
A trimmed name (character scalar).
}
\description{
Validate a field name (internal)
}
\keyword{internal}
