% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-validation.R
\name{.validate_return_FUN}
\alias{.validate_return_FUN}
\title{Validate a return_FUN}
\usage{
.validate_return_FUN(return_FUN)
}
\arguments{
\item{return_FUN}{A function}
}
\value{
The validated \code{return_FUN} (invisibly).
}
\description{
check whether a return FUN used in \code{value_of} and \code{value_of_log}
is a function.
}
\keyword{internal}
