% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_observers.R
\name{extract_observers}
\alias{extract_observers}
\title{Extract data for one or multiple observers}
\usage{
extract_observers(observer_numbers, start_date = "2019-01-01",
  end_date = "2019-02-01", region_type = "country",
  region_id = "kenya", return_type = "data", return_format = "CSV")
}
\arguments{
\item{observer_numbers}{A character or a vector of characters that represent the ADU number or numbers that you want to query the database for. Example: observer_numbers = '10723' or observer_numbers = c('10723', '40147')}

\item{start_date}{A character representing the start date from which you want to query the database. It must be in the YYYY-MM-DD format. Defaults to 1 Jan,2019. Example: '2019-01-26'.}

\item{end_date}{A character representing the end date to which you want to query the database. It must be in the YYYY-MM-DD format. Defaults to 1 Feb, 2019. Example: '2019-02-26'.}

\item{region_type}{Can take one of the following types: country, pentad, group, qdgc. Default is 'country'}

\item{region_id}{Has to be co-defined with `region_type`. That is, it is an instantiation of the `region_type`. For instance if the `region_type` = 'country', then the `region_id` = 'kenya'. Or if `region_type` = 'pentad', then  `region_id` = '3355_1825'. Default is 'kenya'.}

\item{return_type}{There are two options, 'data' and 'count'. 'data' is the actual records submitted to the africa bird map. 'count' return a count of the rows. Default is 'data'.}

\item{return_format}{Default s 'CSV', though the API default is 'JSON'}
}
\value{
Depending on the return_format, the function will return a dataframe with the following:
1. `return_type = data` A tidy dataframe is returned with the records that fall within the specification defined.
2. `return_type = count` A tidy dataframe is returns with the number of rows that will be produced if the return_type was to be specified as 'data'.
}
\description{
Extract data for one or multiple observers
}
\examples{

\dontrun{

# Extract data with single observers.

extract_observers(
observer_numbers = 10723,
start_date = '2019-01-01',
end_date = '2019-02-01',
region_type = 'country',
region_id = 'kenya',
return_type = 'data',
return_format = 'CSV'
)


# Extract data for multiple observers
extract_observers(observer_numbers = c('10723', '40147'),
start_date = '2019-01-01',
end_date = '2019-02-01',
region_type = 'country',
region_id = 'kenya',
return_type = 'data',
return_format = 'CSV'
)


}


}
