% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_species.R
\name{find_species}
\alias{find_species}
\title{Find a specific species or a group of species from a species list}
\usage{
find_species(species_list, species_names)
}
\arguments{
\item{species_list}{The output of `get_species_list`}

\item{species_names}{A character or a vector of characters to search for in a species list,}
}
\value{
A subset of the output of `species_list` that produces the best match for the `species_names` provided.
}
\description{
Find a specific species or a group of species from a species list
}
\examples{

# get species list for kenya
kenya_list <- get_species_list("kenya")

# specify the species you want to search for
interesting_species <- c("drongo", "sunbird")

# find interesting species in the species list
find_species(species_list = kenya_list, species_name = interesting_species)


}
