% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_list.R
\name{get_species_list}
\alias{get_species_list}
\title{Get species list for a country}
\usage{
get_species_list(country)
}
\arguments{
\item{country}{Full name for country}
}
\value{
Dataframe with the species ids, common names and scientific names of species in the defined country. Not when a species is split, the old species ID is discontinued and two (or more) new species ids are created.
}
\description{
Get species list for a country
}
\examples{

# get species list for Nigerian

get_species_list(country = "Nigeria")


}
