% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators-aid.R
\name{\%aid\%}
\alias{\%aid\%}
\alias{\%aid_dp\%}
\alias{\%aid\%<-}
\title{Agent Selection and Replacement by Agents IDs}
\usage{
agents \%aid\% ID

agents \%aid_dp\% ID

`\%aid\%`(agents, ID) <- value
}
\arguments{
\item{agents}{A list of ABM_Agent objects.}

\item{ID}{An integer vector of ID(s) to extract or replace.}

\item{value}{A list of replacement agents. Must match the length of \code{ID}.}
}
\value{
\itemize{
\item \verb{\%aid\%} returns a list of agents which are shallow copied.
\item \verb{\%aid_dp\%} returns a list of agents which are deep copied.
\item \verb{\%aid\%<-} returns the modified agent list.
}
}
\description{
These operators enable efficient and intuitive extraction and replacement of
agents from either an \strong{\code{ABM_G}} object (via its \verb{$agents} list) or
from a standalone list of \strong{\code{ABM_Agent}} objects.

The \verb{\%aid\%} operator extracts agents by ID and returns shallow references
to the agents (i.e., original objects). In contrast, the \verb{\%aid_dp\%} operator
returns deep clones of the agents instead.

The \verb{\%aid\%<-} replacement operator updates the agents at the given IDs.
The length of \code{value} must match the length of \code{ID}.

If any of the provided IDs are not found among the agents, a warning is issued
and those IDs are skipped. The corresponding value entries are also
skipped accordingly.
}
\examples{
# Create and extract agents
agents <- init_agents(n = 3, attr_df = data.frame(x = 1:3), ID_start = 2) # agent's ID starts from ID2
agents \%aid\% c(2, 4)

# Use deep clone if needed
agents \%aid\% c(2, 4)                  # This will return shallow copies
agents \%aid_dp\% c(2, 4)                # This will return deep copies

# # Replace the agent with ID 2 with the agent with ID 3
agents \%aid\% 2 <- agents \%aid\% 3

# Use within ABM_G object
agents <- init_agents(n = 3, attr_df = data.frame(x = 1:3), ID_start = 2) # agent's ID starts from ID2
G <- setABM(agents = agents)
G$agents \%aid\% c(2, 3)


}
\seealso{
\code{\link{\%ai\%}} for accessing agents by index, and
\code{\link{\%aa\%}} for accessing and modifying agent attributes.
}
