% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qap_terms.R
\name{qap_matrix_terms}
\alias{qap_matrix_terms}
\alias{qap_popularity}
\alias{qap_activity}
\alias{qap_reciprocity}
\alias{qap_transitivity}
\alias{qap_common_source}
\alias{qap_common_target}
\title{QAP terms derived from a matrix}
\usage{
qap_popularity(data)

qap_activity(data)

qap_reciprocity(data)

qap_transitivity(data)

qap_common_source(data)

qap_common_target(data)
}
\arguments{
\item{data}{A matrix, typically an adjacency matrix.}
}
\value{
A matrix.
}
\description{
These functions generate matrix-valued QAP terms from an input matrix.
}
\details{
\code{qap_popularity()} returns a matrix in which each row repeats the column sums
of \code{data}.

\code{qap_activity()} returns a matrix in which each column repeats the row sums
of \code{data}.

\code{qap_reciprocity()} returns the transpose of \code{data}.

\code{qap_transitivity()} returns the matrix product \code{data \%*\% data}.

\code{qap_common_source()} returns \code{t(data) \%*\% data}.

\code{qap_common_target()} returns \code{data \%*\% t(data)}.
}
\examples{
mat <- matrix(
  c(0, 1, 1,
    1, 0, 0,
    0, 1, 0),
  nrow = 3, byrow = TRUE
)

qap_popularity(mat)
qap_activity(mat)
qap_reciprocity(mat)
qap_transitivity(mat)
qap_common_source(mat)
qap_common_target(mat)

}
