% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-chunk_modify.R
\name{remove_chunk}
\alias{remove_chunk}
\title{Remove Chunks from a Series Object}
\usage{
remove_chunk(S, ...)
}
\arguments{
\item{S}{A \code{Series} object.}

\item{...}{One or more character strings specifying the names of the
chunks to remove.}
}
\value{
A \code{Series} object with the specified chunks removed.
}
\description{
Removes one or more \code{Chunk} objects from an existing \code{Series}
object by name.
}
\details{
Since \code{Series} is an S3 object, this function returns a modified
copy. The result must be assigned to take effect:
\code{S <- remove_chunk(S, ...)}.
}
\examples{
s1 <- Chunk({ x <- 1 })
s2 <- Chunk({ y <- 2 })
S <- Series(s1, s2)

# remove a single chunk
S <- remove_chunk(S, "s1")

# remove multiple chunks
S <- remove_chunk(S, "s1", "s2")

}
\seealso{
\code{\link{add_chunk}}, \code{\link{replace_chunk}}
}
