% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-chunk_modify.R
\name{replace_chunk}
\alias{replace_chunk}
\title{Replace Chunks in a Series Object}
\usage{
replace_chunk(S, ...)
}
\arguments{
\item{S}{A \code{Series} object.}

\item{...}{One or more \code{Chunk} objects to replace. Each chunk must
have a name that matches an existing chunk in \code{S}.}
}
\value{
A \code{Series} object with the specified chunks replaced.
}
\description{
Replaces one or more \code{Chunk} objects in an existing \code{Series}
object. The names of the new chunks must match those of the chunks to
be replaced.
}
\details{
Since \code{Series} is an S3 object, this function returns a modified
copy. The result must be assigned to take effect:
\code{S <- replace_chunk(S, ...)}.
}
\examples{
s1 <- Chunk({ x <- 1 })
s2 <- Chunk({ y <- 2 })
S <- Series(s1, s2)

# replace chunks
s1 <- Chunk({ x <- 10 })
s2 <- Chunk({ y <- 20 })
S2 <- replace_chunk(S, s1, s2)

}
\seealso{
\code{\link{add_chunk}}, \code{\link{remove_chunk}}
}
