% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_obj.R
\name{copy_obj}
\alias{copy_obj}
\alias{copy_obj.ABM_Game}
\alias{copy_obj.R6}
\title{Deep copy ABM and R6 objects}
\usage{
copy_obj(x)

copy_obj(x)

\method{copy_obj}{ABM_Game}(x)

\method{copy_obj}{R6}(x)
}
\arguments{
\item{x}{An object to copy. Supported classes are:
\itemize{
\item \code{ABM_Agent}
\item \code{ABM_Group}
\item \code{ABM_Game}
\item \code{R6}
}}
}
\value{
A deep-cloned object of the same class as \code{x}.
}
\description{
Create a deep copy of an ABM or R6 object.
}
\details{
This function provides a unified interface for cloning objects used in
rABM. For ABM and R6 objects, it performs a deep clone using
\code{$clone(deep = TRUE)}.

For \code{ABM_Agent}, \code{ABM_Game}, and other \code{R6} objects,
\code{copy_obj()} calls \code{$clone(deep = TRUE)}.

For \code{ABM_Group}, which is internally represented as a list of agents,
each agent is deep-cloned individually, and the group structure (including
element names) is preserved.

Objects that do not belong to the supported classes are not handled by
this function and will result in an error.
}
\examples{
# Pseudo-code (depends on constructors):
# A  <- Agent(state = list(a = 1))
# A2 <- copy_obj(A)

# Gr  <- Group(list(ID1 = A, ID2 = A))
# Gr2 <- copy_obj(Gr)

# G  <- Game(stage = list(x = 1), group = Gr)
# G2 <- copy_obj(G)
}
