% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-format_input.R
\name{.format_input}
\alias{.format_input}
\title{Normalize user input into a named list}
\usage{
.format_input(x, expr, prefix = "Z")
}
\arguments{
\item{x}{An input object to be normalized.}

\item{expr}{The unevaluated expression corresponding to \code{x}
(typically obtained via \code{match.call()}).}

\item{prefix}{A prefix used to generate names when automatic naming is required.}
}
\value{
A named list.
}
\description{
Internal utility function used to normalize user inputs for API functions
such as \code{\link[=Game]{Game()}}.
}
\details{
\itemize{
\item A single object is wrapped into a list.
\item Unnamed single objects may have their name inferred from the call.
\item Unnamed multiple inputs are automatically named using a prefix.
}
}
\keyword{internal}
