% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-plot-Game.R
\name{plot.ABM_Game}
\alias{plot.ABM_Game}
\alias{plot_Game}
\title{Plot an ABM_Game object (current state or log replay)}
\usage{
\method{plot}{ABM_Game}(x, name = NULL, log = NULL, verbose = TRUE, ...)

plot_Game(G, name = NULL, log = NULL, ...)
}
\arguments{
\item{x}{An \code{ABM_Game} object \code{G}.}

\item{name}{Name of a \code{plot_FUN} field in \code{x}. If \code{NULL}, all
fields categorized as \code{"plot_FUN"} are executed sequentially.}

\item{log}{Which log entries to replay. If \code{NULL}, the plot is produced
using the current state of \code{x}. If provided, it must be either a
character vector matching \code{names(x$log)} or a numeric vector of indices
within \code{seq_len(length(x$log))}.}

\item{...}{Additional arguments passed to the underlying \code{plot_FUN}.}
}
\value{
\code{NULL} invisibly.
}
\description{
Execute one or more \code{plot_FUN} functions stored in an \code{ABM_Game}
object. You can either plot the current state (\code{log = NULL}) or replay
plots over stored snapshots in \code{x$log}.
}
\details{
See \code{plot.ABM_Game()} for details.
}
\examples{
\dontrun{
plot(G)
plot(G, name = "plot_map")
plot(G, name = "plot_map", log = 1:5)
plot(G, name = "plot_map", log = c("t1", "t5"))
}
}
