% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rABM-package.R
\docType{package}
\name{rABM-package}
\alias{rABM}
\alias{rABM-package}
\title{rABM: A flexible and extensible agent-based modeling package for exploratory social science research}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Keiichi Satoh \email{keiichi.satoh@r.hit-u.ac.jp}

}
\keyword{internal}
